+++ src/frontend/apps/desk/src/features/teams/api/index.ts
@@ -2,5 +2,3 @@ export * from './types';
 export * from './useCreateTeam';
 export * from './useTeam';
 export * from './useTeams';
-export * from './useTeamsAccesses';
-export * from './useUpdateTeamAccess';
